+++ src/main/java/study/posthub/domain/member/controller/AuthController.java
@@ -8,11 +8,10 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.bind.annotation.SessionAttribute;
 import study.posthub.domain.member.entity.Member;
 import study.posthub.domain.member.service.MemberService;
-import study.posthub.global.security.LoginUser;
+import study.posthub.global.security.oauth2.LoginMember;
+import study.posthub.global.security.oauth2.SessionMember;
 
 import java.io.IOException;
 
@@ -26,13 +25,14 @@ public class AuthController {
     private final MemberService memberService;
 
     @GetMapping("/")
-    public String mainPage(Model model, @LoginUser Member member, Long userId) {
+    public String mainPage(Model model, @LoginMember SessionMember member, Long userId) {
 //        Member user = memberService.getLoginUser(userId);
-        log.info("member: {}", member);
 
         if (member == null) {
             member = createAnonymousMember();
         }
+
+        log.info("member: {}", member);
         model.addAttribute("user", member);
         return "main";
     }
@@ -54,10 +54,10 @@ public class AuthController {
         //response.sendRedirect("https://nid.naver.com/nidlogin.logout?returl="); //위 코드대신 써도됨 써도 됨.
     }
 
-    private Member createAnonymousMember() {
+    private SessionMember createAnonymousMember() {
         Member anonymousMember = new Member();
         anonymousMember.setNickname("익명");
         anonymousMember.setAuthority(SIGNOUT);
-        return anonymousMember;
+        return new SessionMember(anonymousMember);
     }
 }
