+++ src/main/java/study/posthub/global/config/WebConfig.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import study.posthub.global.security.LoginUserArgumentResolver;
+import study.posthub.global.security.oauth2.LoginUserArgumentResolver;
 
 import java.util.List;
