+++ src/main/java/study/posthub/global/security/oauth2/LoginMember.java
@@ -1,4 +1,4 @@
-package study.posthub.global.security;
+package study.posthub.global.security.oauth2;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,5 +7,5 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.PARAMETER)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface LoginUser {
+public @interface LoginMember {
 }
