+++ src/main/java/study/posthub/global/security/oauth2/NaverResponse.java
@@ -1,27 +1,23 @@
-package study.posthub.domain.member.dto.impl;
-
-import lombok.RequiredArgsConstructor;
-import study.posthub.domain.member.dto.OAuth2Response;
+package study.posthub.global.security.oauth2;
 
 import java.util.Map;
 
-@RequiredArgsConstructor
-public class GoogleResponse implements OAuth2Response {
+public class NaverResponse implements OAuth2Response {
 
     private final Map<String, Object> attribute;
 
-    public GoogleResponse(Map<String, Object> attribute) {
-        this.attribute = attribute;
+    public NaverResponse(Map<String, Object> attribute) {
+        this.attribute = (Map<String, Object>) attribute.get("response");
     }
 
     @Override
     public String getProvider() {
-        return "google";
+        return "naver";
     }
 
     @Override
     public String getProviderId() {
-        return attribute.get("sub").toString();
+        return attribute.get("id").toString();
     }
 
     @Override
