+++ src/main/java/study/posthub/global/security/oauth2/SessionMember.java
@@ -1,20 +1,25 @@
-package study.posthub.global.security;
+package study.posthub.global.security.oauth2;
 
 import lombok.Getter;
+import lombok.ToString;
+import study.posthub.domain.member.entity.Authority;
 import study.posthub.domain.member.entity.Member;
 
 import java.io.Serializable;
 
 @Getter
-public class SessionUser implements Serializable {
+@ToString
+public class SessionMember {
 
     private Long id;
     private String email;
     private String nickname;
+    private Authority authority;
 
-    public SessionUser(Member member) {
+    public SessionMember(Member member) {
         this.id = member.getId();
         this.email = member.getEmail();
         this.nickname = member.getNickname();
+        this.authority = member.getAuthority();
     }
 }
